refactor(cubejs-testing): move index granularity tests into smoke-postgres
Merge the index granularity E2E tests into the existing
smoke-postgres.test.ts instead of having a separate test file.
Both share the same Birdbox setup (smoke/schema dir, smoke/cube.js).
