+++ packages/cubejs-testing/package.json
@@ -78,7 +78,6 @@
     "smoke:multidb": "jest --verbose -i dist/test/smoke-multidb.test.js",
     "smoke:multidb:snapshot": "jest --verbose --updateSnapshot -i dist/test/smoke-multidb.test.js",
     "smoke:postgres": "jest --verbose -i dist/test/smoke-postgres.test.js",
-    "smoke:index-granularity": "jest --verbose -i dist/test/smoke-index-granularity.test.js",
     "smoke:redshift": "jest --verbose -i dist/test/smoke-redshift.test.js",
     "smoke:redshift:snapshot": "jest --verbose --updateSnapshot -i dist/test/smoke-redshift.test.js",
     "smoke:vertica": "jest --verbose -i dist/test/smoke-vertica.test.js",
